requirements.txt
django
pillow
psycopg2-binary
djangorestframework
